# Conflicts:
#	src/Nethermind/Nethermind.Core/Specs/IReleaseSpec.cs
#	src/Nethermind/Nethermind.JsonRpc.Test/Modules/Eth/EthRpcModuleTests.cs
#	src/Nethermind/Nethermind.Specs.Test/OverridableReleaseSpec.cs
#	src/Nethermind/Nethermind.Specs/ChainSpecStyle/ChainParameters.cs
#	src/Nethermind/Nethermind.Specs/ChainSpecStyle/ChainSpecBasedSpecProvider.cs
#	src/Nethermind/Nethermind.Specs/Forks/00_Olympic.cs
#	src/Nethermind/Nethermind.Specs/Forks/01_Frontier.cs
#	src/Nethermind/Nethermind.Specs/Forks/02_Homestead.cs
#	src/Nethermind/Nethermind.Specs/Forks/03_Dao.cs
#	src/Nethermind/Nethermind.Specs/Forks/04_TangerineWhistle.cs
#	src/Nethermind/Nethermind.Specs/Forks/05_SpuriousDragon.cs
#	src/Nethermind/Nethermind.Specs/Forks/06_Byzantium.cs
#	src/Nethermind/Nethermind.Specs/Forks/07_Constantinople.cs
#	src/Nethermind/Nethermind.Specs/Forks/08_ConstantinopleFix.cs
#	src/Nethermind/Nethermind.Specs/Forks/09_Istanbul.cs
#	src/Nethermind/Nethermind.Specs/Forks/10_MuirGlacier.cs
#	src/Nethermind/Nethermind.Specs/Forks/11_Berlin.cs
#	src/Nethermind/Nethermind.Specs/Forks/12_London.cs
#	src/Nethermind/Nethermind.Specs/SystemTransactionReleaseSpec.cs

+++ src/Nethermind/Nethermind.JsonRpc/Modules/Trace/ITraceRpcModule.cs
@@ -54,7 +54,7 @@ namespace Nethermind.JsonRpc.Modules.Trace
         ResultWrapper<ParityTxTraceFromStore[]> trace_block([JsonRpcParameter(ExampleValue = "latest")] BlockParameter numberOrTag);
         
         [JsonRpcMethod(Description = "", IsImplemented = false, IsSharable = false)]
-        ResultWrapper<ParityTxTraceFromStore[]> trace_get(Keccak txHash, int[] positions);
+        ResultWrapper<ParityTxTraceFromStore[]> trace_get(Keccak txHash, long[] positions);
         
         [JsonRpcMethod(Description = "", IsImplemented = true, 
             IsSharable = false,
